API Reference

Main Functions

Speasy.get_data — Function

get_data(args...; drop_nan=false)

Get data using speasy Python package. We support the same arguments as speasy.get_data.

Set drop_nan=true to drop the nan values. Note that we need to do that in Python since we cannot convert NaT (not a time) to Julia.

Speasy.@spz_str — Macro

spz"provider/dataset/parameter"
spz"provider/dataset/parameter1,parameter2"

String macro to create a SpeasyProduct from a string identifier. Supports multiple parameters separated by commas, which returns a tuple of SpeasyProduct objects.

Examples

# Single parameter
product = spz"cda/OMNI_HRO_1MIN/flow_speed"

# Multiple parameters
products = spz"cda/OMNI_HRO_1MIN/flow_speed,Pressure"

Speasy.list_parameters — Function

list_parameters(provider, [dataset]; verbose=false)

Find the available parameters for a given provider or for a specific dataset from provider.

Set verbose=true to print the metadata of the dataset.

Examples

# List all parameters from AMDA provider
list_parameters(:amda)

# List parameters from specific CDA dataset
list_parameters(:cda, "SOHO_ERNE-HED_L2-1MIN")

# output
5-element Vector{String}:
 "est"
 "PH"
 "AH"
 "PHC"
 "AHC"

See also: find_datasets

Speasy.find_datasets — Function

find_datasets(provider, [term...])

Find the available datasets for a given provider, optionally filtered by search terms (only datasets containing all specified terms will be returned.)

Examples

# List all datasets from AMDA provider
find_datasets(:amda)

# List CDA datasets containing "OMNI"
find_datasets(:cda, :OMNI)

# List CDA datasets containing both "OMNI" and "HRO"
find_datasets(:cda, :OMNI, :HRO)

# output
4-element Vector{String}:
 "OMNI_HRO_1MIN"
 "OMNI_HRO2_1MIN"
 "OMNI_HRO_5MIN"
 "OMNI_HRO2_5MIN"

See also: list_parameters

Data Types

Speasy.SpeasyVariable — Type

A wrapper of speasy.SpeasyVariable.

Speasy.VariableAxis — Type

A wrapper of speasy.VariableAxis. https://github.com/SciQLop/speasy/blob/main/speasy/core/data_containers.py#L234

Data Access and Processing

Speasy.sanitize! — Function

Replaces invalid values and fill values by NaN for var with float type elements.

Speasy.replace_invalid! — Function

Replaces invalid values by NaN for var with float type elements.

Speasy.replace_fillval_by_nan! — Function

Replaces fill values by NaN for var with float type elements.

Public

Speasy.Speasy — Module

A Julia wrapper around speasy, a Python package to deal with main Space Physics WebServices.

Space Physics made EASY!

Links: GitHub, Documentation

Speasy.SpeasyVariable — Type

A wrapper of speasy.SpeasyVariable.

Speasy.VariableAxis — Type

A wrapper of speasy.VariableAxis. https://github.com/SciQLop/speasy/blob/main/speasy/core/data_containers.py#L234

Speasy.find_datasets — Method

find_datasets(provider, [term...])

Find the available datasets for a given provider, optionally filtered by search terms (only datasets containing all specified terms will be returned.)

Examples

# List all datasets from AMDA provider
find_datasets(:amda)

# List CDA datasets containing "OMNI"
find_datasets(:cda, :OMNI)

# List CDA datasets containing both "OMNI" and "HRO"
find_datasets(:cda, :OMNI, :HRO)

# output
4-element Vector{String}:
 "OMNI_HRO_1MIN"
 "OMNI_HRO2_1MIN"
 "OMNI_HRO_5MIN"
 "OMNI_HRO2_5MIN"

See also: list_parameters

Speasy.get_data — Method

get_data(args...; drop_nan=false)

Get data using speasy Python package. We support the same arguments as speasy.get_data.

Set drop_nan=true to drop the nan values. Note that we need to do that in Python since we cannot convert NaT (not a time) to Julia.

Speasy.list_parameters — Method

list_parameters(provider, [dataset]; verbose=false)

Find the available parameters for a given provider or for a specific dataset from provider.

Set verbose=true to print the metadata of the dataset.

Examples

# List all parameters from AMDA provider
list_parameters(:amda)

# List parameters from specific CDA dataset
list_parameters(:cda, "SOHO_ERNE-HED_L2-1MIN")

# output
5-element Vector{String}:
 "est"
 "PH"
 "AH"
 "PHC"
 "AHC"

See also: find_datasets

Speasy.replace_fillval_by_nan! — Method

Replaces fill values by NaN for var with float type elements.

Speasy.replace_invalid! — Method

Replaces invalid values by NaN for var with float type elements.

Speasy.sanitize! — Method

Replaces invalid values and fill values by NaN for var with float type elements.

Speasy.@spz_str — Macro

spz"provider/dataset/parameter"
spz"provider/dataset/parameter1,parameter2"

String macro to create a SpeasyProduct from a string identifier. Supports multiple parameters separated by commas, which returns a tuple of SpeasyProduct objects.

Examples

# Single parameter
product = spz"cda/OMNI_HRO_1MIN/flow_speed"

# Multiple parameters
products = spz"cda/OMNI_HRO_1MIN/flow_speed,Pressure"